+++ packages/ui/src/components/message/components/document-copy/index.tsx
@@ -5,6 +5,7 @@ import {
   MessageDocumentContentBlockTop,
   MessageDocumentCopyButton,
 } from './styled';
+import { processSelection } from '@/ui/utils';
 
 interface GeneratedDocumentBlockProps {
   code: string;
@@ -29,48 +30,24 @@ export const GeneratedDocumentBlock = ({
   const handleCopy = async () => {
     if (!contentRef.current) return;
 
-    const selection = window.getSelection();
-    const originalRange =
-      selection && selection.rangeCount > 0 ? selection.getRangeAt(0) : null;
-
     const range = document.createRange();
     range.selectNodeContents(contentRef.current);
-    selection?.removeAllRanges();
-    selection?.addRange(range);
 
-    const dt = new DataTransfer();
-    const copyEvent = new ClipboardEvent('copy', {
-      clipboardData: dt,
-      bubbles: true,
-      cancelable: true,
-    });
+    const { html, plain } = processSelection(range, !!isTg, !isTg);
 
     try {
-      contentRef.current.dispatchEvent(copyEvent);
-    } finally {
-      selection?.removeAllRanges();
-      if (originalRange) {
-        selection?.addRange(originalRange);
+      const clipboardItems: Record<string, Blob> = {};
+
+      if (plain) {
+        clipboardItems['text/plain'] = new Blob([plain], {
+          type: 'text/plain',
+        });
       }
-    }
-
-    try {
-      const html = dt.getData('text/html');
-      const plain = dt.getData('text/plain');
-
-      if (html || plain) {
-        const clipboardItems: Record<string, Blob> = {};
-        if (plain)
-          clipboardItems['text/plain'] = new Blob([plain], {
-            type: 'text/plain',
-          });
-        if (html)
-          clipboardItems['text/html'] = new Blob([html], { type: 'text/html' });
-
-        await navigator.clipboard.write([new ClipboardItem(clipboardItems)]);
-      } else {
-        await navigator.clipboard.writeText(contentRef.current.innerText);
+      if (html) {
+        clipboardItems['text/html'] = new Blob([html], { type: 'text/html' });
       }
+
+      await navigator.clipboard.write([new ClipboardItem(clipboardItems)]);
 
       setCopied(true);
       setTimeout(() => setCopied(false), 2000);
